% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_models.R
\name{bs_diff}
\alias{bs_diff}
\title{B spline basis with option of centering and differencing coefficients}
\usage{
bs_diff(
  x,
  df = NULL,
  knots = NULL,
  degree = 3,
  intercept = TRUE,
  Boundary.knots = range(x),
  differences = 1,
  periodic = FALSE,
  center = TRUE
)
}
\value{
list of data_model variables including: \itemize{
    \item state a list of parameters associated with the data_model. Here, only the matrix Eta
    \item posteriorSample_params a list of parameter names to record posterior samples of
    \item posteriorMean_params a list of parameters to record the posterior mean, but not save individual
    posterior samples
}
}
\description{
parameters follow \link{bs}
}
\details{
uses code from \link{bs}, but optionally transforms the basis into a difference
between spline parameters.

Also, optionally centers the spline so that the spline average is zero.
}
\references{
following code from https://github.com/SurajGupta/r-source/blob/master/src/library/splines/R/splines.R
}
