% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_models.R
\name{cis_eQTL_model}
\alias{cis_eQTL_model}
\title{Sample cis_eQTL coefficients}
\usage{
cis_eQTL_model(observation_model_parameters, MegaLMM_state = list())
}
\arguments{
\item{observation_model_parameters}{list with:
\code{Y} gene expression data matrix n x p
\code{cis_genotypes} a list of design matrices of length \code{p} (ie number of columns of \code{Eta})
This is used to specify trait-specific fixed effects, such a cis-genotypes}

\item{MegaLMM_state}{}
}
\value{
list including:
    \code{state} parameters to add to \code{current_state}
    \code{posteriorSample_params} character vector of parameter names to include in the list of parameters to record posterior samples
    \code{posteriorMean_params} character vector of parameter names to include in the list of parameters to record posterior mean
}
\description{
Sample cis_eQTL coefficients
}
