% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_functions.R
\name{clear_Posterior}
\alias{clear_Posterior}
\title{Resets Posterior samples}
\usage{
clear_Posterior(MegaLMM_state)
}
\description{
Clears and resets the saved Posterior samples. Updates the burn parameter of
    \code{run_parameters} to reflect all previous samples in the chain now count as burnin
}
\seealso{
\code{\link{sample_MegaLMM}}, \code{\link{plot.MegaLMM_state}}
}
