% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_models.R
\name{combined_model}
\alias{combined_model}
\title{Sample Eta given a list of different observation_models}
\usage{
combined_model(observation_model_parameters, MegaLMM_state = list())
}
\arguments{
\item{observation_model_parameters}{a list of observation_model lists}

\item{MegaLMM_state}{the current MegaLMM_state object. Can be just a list with \code{data}}
}
\value{
list of data_model variables including:

state a list of parameters associated with the data_model. Here, only the matrix Eta

posteriorSample_params a list of parameter names to record posterior samples of

posteriorMean_params a list of parameters to record the posterior mean, but not save individual
    posterior samples
}
\description{
Sample Eta given a list of different observation_models
}
