% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_data_matrices}
\alias{create_data_matrices}
\title{Create data matrices for MegaLMM from a "tall" data.frame}
\usage{
create_data_matrices(tall_data, id_cols, names_from, values_from, ...)
}
\arguments{
\item{tall_data}{data.frame in the tall format}

\item{id_cols}{vector of strings representing the set of columns of \code{tall_data} 
necessary to uniquely specify a single individual. See \code{\link{tidyr::pivot_wider}}.}

\item{names_from}{vector of strings representing the set of columns of \code{tall_data} 
necessary to uniquely specify a single trait See \code{\link{tidyr::pivot_wider}}.}

\item{values_from}{string giving the column of \code{tall_data} with the observations.}

\item{...}{additional parameters passed to \code{\link{tidyr::pivot_wider}}.}
}
\value{
list with elements:
* \code{data} a data.frame with the elements of \code{id_cols}
* \code{Y} a numeric matrix of the traits
* \code{trait_info} a data.frame with 1 row per trait and columns from \code{names_from}
}
\description{
Often multi-trait data will be provided in tall format with a single value
per row, and individuals represented on multiple rows. MegaLMM expects the 
data to be in wide format with all observations on the same individual in a row,
and only one row per individual.
}
\details{
Specifically, the input to \code{\link{setup_model_MegaLMM}} expect \code{Y}, a 
\code{nxp} matrix for \code{n} individuals and \code{p} traits (Missing values coded as \code{NA}
are fine in \code{Y}), and \code{data}, a data.frame with the variables used in \code{formula}.
This function takes the tall format data and pivots it using the \code{\link{tidyr::pivot_wider}}
function, returning the \code{Y} and \code{data} objects necessary to set up a
MegaLMM model
}
