% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{estimate_memory_initialization_MegaLMM}
\alias{estimate_memory_initialization_MegaLMM}
\title{Estimate the memory requirment for a fully initialized MegaLMM model}
\usage{
estimate_memory_initialization_MegaLMM(MegaLMM_state)
}
\arguments{
\item{MegaLMM_state}{The model after calling \code{initialize_variables_MegaLMM}}
}
\value{
The estimated memory size in bytes
}
\description{
the \code{initialize_MegaLMM()} function can take a while to run, especially
when there are multiple random effects, large numbers of observations, and multiple
groups of traits (with different missing data patterns, stored in the \code{Missing_data_map}).
Approximately, the initialization time and memory requirements will be linear in the 
number of groups of traits, and scale with h2_divisions^(# random effects).
}
\details{
This function will initialize the MegaLMM model for a single h2 vector and extrapolate
to the full grid of h2 vectors, enabeling you to estimate if you have enough
memory allocated to call \code{initialize_MegaLMM()}
}
\examples{
estimate_memory_initialization_MegaLMM(MegaLMM_state)
}
\seealso{
\code{\link{estimate_memory_posterior}}
}
