% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_candidate_states}
\alias{find_candidate_states}
\title{Finds the set of variance component proportions within a specified distance from a starting proportion}
\usage{
find_candidate_states(h2s_matrix, step_size, old_state)
}
\arguments{
\item{h2s_matrix}{Mxl matrix of all valid variance component proportions for M random effects}

\item{step_size}{value measuring the maximum euclidean distance to a new set of variance component proportions}

\item{old_state}{index in h2s_matrix of the current value of the variance component proportions}
}
\value{
vector of indices of \code{h2s_matrix} giving new candidate variance component proportions
}
\description{
Finds the set of variance component proportions within a specified distance from a starting proportion
}
