% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_functions.R
\name{get_posterior_HPDinterval}
\alias{get_posterior_HPDinterval}
\title{Calculates Highest Posteriod Density intervals of a function of parameters}
\usage{
get_posterior_HPDinterval(X, FUN = NULL, prob = 0.95, ...)
}
\arguments{
\item{X}{either an array of posterior samples (either a parameter from \code{Posterior}, or
an object generated by \link{get_posterior_FUN}), or the MegaLMM_state object with re-loaded Posterior}

\item{FUN}{(optional) if \code{X} is a MegaLMM_state object, the function to calculate the synthetic parameter}

\item{prob}{A numeric scalar in the interval (0,1) giving the target probability content of the intervals.
The nominal probability content of the intervals is the multiple of 1/nrow(obj) nearest to prob}

\item{...}{other parameters passed to \link{get_posterior_FUN}}
}
\value{
array with first dimension of length=2 giving the lower and upper limits of the parameters
    in the other two dimensions.
}
\description{
The synthetic parameter can be pre-calculated using \link{get_posterior_FUN}, or provided directly
    as an array with the first dimension the samples.#'
}
