% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_functions.R
\name{get_posterior_mean}
\alias{get_posterior_mean}
\title{Calculates posterior mean of a function of parameters}
\usage{
get_posterior_mean(X, FUN, bychunk = FALSE, mc.cores = 1, ...)
}
\arguments{
\item{X}{either an array of posterior samples (either a parameter from \code{Posterior}, or
an object generated by \link{get_posterior_fun}), or the MegaLMM_state object with re-loaded Posterior}

\item{FUN}{(optional) if \code{X} is a MegaLMM_state object, the function to calculate the synthetic parameter}

\item{bychunk}{(optional) if \code{TRUE}, will re-load each chunk of the Posterior separately
and calculate the posterior mean of \code{FUN} separately for each chunk, and then return
the overall posterior mean. This saves memory because the whole posterior does not need to be
loaded into memory at once. Only necessary terms from Posterior are loaded.}
}
\value{
posterior mean matrix
}
\description{
The synthetic parameter can be pre-calculated using \link{get_posterior_FUN}, or provided directly
    as an array with the first dimension the samples.
}
