% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_functions.R
\name{Image}
\alias{Image}
\title{Modification to \code{image()}.}
\usage{
Image(X, dimnames = FALSE, title = NULL, include_zero = TRUE, ...)
}
\arguments{
\item{X}{numeric matrix to plot.}

\item{dimnames}{should row/colnames of matrix be added to the plot}

\item{title}{Optional title}

\item{include_zero}{should the z-value zero be included in the scale?}
}
\value{
ggplot object
}
\description{
Like `image,CHMfactor-method`, but with better defaults for correlation matrices especially.
}
\details{
The default is to have the colors evenly spaced, instead of having the full range <0 and >0.
}
