% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_master.R
\name{initialize_MegaLMM}
\alias{initialize_MegaLMM}
\title{Initialized Gibbs sampler for MegaLMM model}
\usage{
initialize_MegaLMM(
  MegaLMM_state,
  ncores = my_detectCores(),
  Qt_list = NULL,
  chol_R_list = NULL,
  chol_ZKZt_list = NULL
)
}
\arguments{
\item{MegaLMM_state}{MegaLMM_state object as returned by \link{setup_model_MegaLMM}}

\item{ncores}{number of cores to use for parallel evaluations. Not really used as RcppParallel is used instead.
Instead, we break up the computation into chunks of this size.}

\item{Qt_list}{Optionally, \code{Qt_list}, \code{chol_R_list} and \code{chol_ZKZt_list} can be provided
from a previous MegaLMM_state object if the data and model is identical.}

\item{chol_R_list}{See \code{Qt_list}}

\item{chol_ZKZt_list}{See \code{Qt_list}}
}
\value{
MegaLMM_state object with \code{Qt_list}, \code{chol_R_list} and \code{chol_ZKZt_list} added to run_variables
}
\description{
The pre-calculates a set of matrices that will be re-used through the Gibbs chains.
These calculations can be slow for large models, especially if n is large, the number of
random effects is > 1, or there are many groups of observations with different
missing data patterns.
}
