% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_master.R
\name{initialize_variables_MegaLMM}
\alias{initialize_variables_MegaLMM}
\title{Initialize MegaLMM variables}
\usage{
initialize_variables_MegaLMM(MegaLMM_state, ...)
}
\arguments{
\item{MegaLMM_state}{MegaLMM_state object as returned by \link{setup_model_MegaLMM}}

\item{...}{Currenlty not supported}
}
\value{
MegaLMM_state object with current_state initialized with all variables
}
\description{
Initializes all variables in MegaLMM model with random draws.
Most variables are drawn from N(0,1) distributions, or uniformly for h2 parameters
}
