% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{K}
\alias{K}
\title{Genomic Relationship matrix of 502 lines}
\format{
A square matrix with row and column names
}
\usage{
K
}
\description{
The genomic relationship matrix is a symmetric, square matrix
giving the average sharing of genotypes across the genome of pairs 
of lines, and used to estimate heritability and do genomic prediction.
This GRM is based on lines used in the GenomesToFields initiative,
but line names have been anonymized.
}
\keyword{datasets}
