% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_multiply_toDense}
\alias{matrix_multiply_toDense}
\title{Multiplies two matrices (sparse or dense by dense), returns the product as a dense matrix}
\usage{
matrix_multiply_toDense(X_, Y_)
}
\arguments{
\item{X_}{First matrix (matrix or dgCMatrix)}

\item{Y_}{Second matrix (matrix)}
}
\value{
Product of X_ and Y_ as a dense matrix
}
\description{
Multiplies two matrices (sparse or dense by dense), returns the product as a dense matrix
}
