% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_master.R
\name{MegaLMM_priors}
\alias{MegaLMM_priors}
\title{Set MegaLMM priors}
\usage{
MegaLMM_priors(tot_Y_var = list(V = 0.5, nu = 3), tot_F_var = list(V =
  18/20, nu = 20), h2_priors_resids_fun = function(h2s, n) 1,
  h2_priors_factors_fun = function(h2s, n) 1,
  Lambda_prior = list(sampler = sample_Lambda_prec_horseshoe, prop_0 =
  0.1, delta_l = list(shape = 3, rate = 1), delta_iterations_factor = 100),
  B2_prior = list(sampler = sample_B2_prec_horseshoe, prop_0 = 0.1),
  cis_effects_prior = list(prec = 1))
}
\arguments{
\item{tot_Y_var}{List of parameters of inverse gamma distribution for residual variances, specifically:
\code{V} and \code{nu}, give shape = \code{nu/2} and scale = \code{nu*V/2}, so mean = \code{(V*nu)/(nu-2)}}

\item{tot_F_var}{List of parameters of inverse gamma distribution for factor variances. See \code{tot_Y_var}.
This parameter provides the parameter extension of Ghosh and Dunson (2009), but is removed
from all Factor parameters before they are saved in Posterior}

\item{h2_priors_resids_fun}{function that returns prior probability for a given vector of variance component proportions for each random effect
The function must take two arguments - a vector of variance component proportions: \code{h2},
and \code{n} - the number of discrete levels of the prior.
Alternatively, can be a scalar or vector of (relative) prior values for each value of the
discrete prior.}

\item{h2_priors_factors_fun}{see \code{h2_priors_resids_fun}. Same, but for the h2s of the factors.}

\item{Lambda_prior}{A list with elements:
1) \code{sampler}: a function that draws samples of the precision matrix for Lambda. Ex: \code{sample_Lambda_prec_reg_horseshoe}; 2)
any other hyperparameters and control parameters for \code{sampler}}

\item{cis_effects_prior}{Currently accepts a list with a single value giving the precision of each cis_effect variable}

\item{B_prior}{A list with elements:
1) \code{sampler}: a function that draws samples of the precision matrix for B and B_F Ex: \code{sample_B2_prec_reg_horseshoe}; 2)
any other hyperparameters and control parameters for \code{sampler}}
}
\value{
a list with each of the prior components specified above.
}
\description{
Function to create list of priors for MegaLMM model.
}
\details{
Default values are provided, but any can be replaced. Note: \code{h2_priors_resids} and
    \code{h2_priors_factors} can be set after calling \code{MegaLMM_init} and before \code{sample_MegaLMM}
    if that is easier. See Vignette.
}
