% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_master.R
\name{plot.MegaLMM_state}
\alias{plot.MegaLMM_state}
\title{Make plots of current MegaLMM state}
\usage{
\method{plot}{MegaLMM_state}(MegaLMM_state,
  file = "diagnostics_plots.pdf", setup = NULL)
}
\arguments{
\item{MegaLMM_state}{output of sample_MegaLMM}

\item{file}{Output file for pdf booklet}

\item{setup}{optional - a list of known values for Lambda (error_factor_lambda), h2, factor_h2s}
}
\description{
Make plots of current MegaLMM state
}
\seealso{
\code{\link{MegaLMM_control}}, \code{\link{sample_MegaLMM}}, \code{\link{MegaLMM_init}},
  \code{\link{print.MegaLMM_state}}, \code{\link{summary.MegaLMM_state}}
}
