% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Posterior.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{Plots diagnostic plots for the fit of a MegaLMM model}
\usage{
plot_diagnostics(MegaLMM_state)
}
\arguments{
\item{MegaLMM_state}{a MegaLMM_state object}
}
\description{
Currently, only traceplots of F_h2 and elements of Lambda are shown
}
