% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_models.R
\name{regression_model}
\alias{regression_model}
\title{Sample Eta given regression-splines individual-level model}
\usage{
regression_model(observation_model_parameters, MegaLMM_state = list())
}
\arguments{
\item{observation_model_parameters}{a list including:
1) \code{observations}, a data.frame with observation-level data including columns \code{ID} and \code{Y}
3) \code{individual_model} the model that should be applied to the data for each ID.}

\item{MegaLMM_state}{The current \code{MegaLMM_state}.
For initialization, can be a list with \code{data_matrices = list(data=data)}
Where \code{data} contains \code{ID} for ordering.}
}
\description{
Eta is a matrix of individual-level parmaters for a regression-spline with
   equivalent knots over all individuals
}
\details{
This function should pre-calculate design matrices for each individual (assuming they are all
    unique). During sampling, should sample regression coefficients \code{Eta} given the factor model state.
    Set up to allow for non iid errors, but currently not implemented.
}
