% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_functions.R
\name{reorder_factors}
\alias{reorder_factors}
\title{Re-orders factors in decreasing order of magnitude}
\usage{
reorder_factors(MegaLMM_state, factor_order = NULL,
  drop_cor_threshold = 0.6)
}
\arguments{
\item{MegaLMM_state}{output of \code{\link{sample_MegaLMM}}}

\item{factor_order}{optional: vector or new orders for the factors. If Null, will be calculated}

\item{drop_cor_threshold}{Factors correlated above this value (either F or U_F) will be dropped}
}
\description{
Re-orders factors in decreasing order of magnitude
}
\seealso{
\code{\link{sample_MegaLMM}}, \code{\link{plot.MegaLMM_state}}
}
