% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_base_samplers.R
\name{sample_MegaLMM}
\alias{sample_MegaLMM}
\title{Run MegaLMM Gibbs sampler}
\usage{
sample_MegaLMM(MegaLMM_state, n_iter, grainSize = 1, verbose = TRUE, ...)
}
\arguments{
\item{MegaLMM_state}{MegaLMM_state object of current chain}

\item{n_iter}{Number of iterations to add to the chain (not number of posterior samples to draw.
This is determined by n_iter / thin)}
}
\description{
Run MCMC chain for a specified number of iterations
}
