% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_functions.R
\name{save_posterior_chunk}
\alias{save_posterior_chunk}
\title{Saves a chunk of posterior samples}
\usage{
save_posterior_chunk(MegaLMM_state)
}
\arguments{
\item{Posterior}{a Posterior list from a MegaLMM_state object}

\item{folder}{the folder to save the RData files}
}
\value{
Posterior
}
\description{
Saves a chunk of posterior samples, each paramter in its own RData file.
The complete chain for a particular parameter can be re-loaded with \link{load_posterior_param}
}
