% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_functions.R
\name{save_posterior_sample}
\alias{save_posterior_sample}
\title{Saves current state in Posterior}
\usage{
save_posterior_sample(MegaLMM_state)
}
\description{
Saves current state in Posterior
}
\seealso{
\code{\link{sample_MegaLMM}}, \code{\link{plot.MegaLMM_state}}
}
