% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_master.R
\name{set_priors_MegaLMM}
\alias{set_priors_MegaLMM}
\title{Set priors for MegaLMM model.}
\usage{
set_priors_MegaLMM(MegaLMM_state, priors = MegaLMM_priors())
}
\arguments{
\item{MegaLMM_state}{MegaLMM_state object as returned by \link{setup_model_MegaLMM}}

\item{priors}{List as returned by \link{MegaLMM_priors}}
}
\value{
MegaLMM_state object with prior information added.
}
\description{
See \link{MegaLMM_priors} for more information
}
