% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MegaLMM_master.R
\name{summary.MegaLMM_state}
\alias{summary.MegaLMM_state}
\title{Print more detailed statistics on current MegaLMM state}
\usage{
summary.MegaLMM_state(MegaLMM_state)
}
\description{
Print more detailed statistics on current MegaLMM state
}
\seealso{
\code{\link{MegaLMM_control}}, \code{\link{sample_MegaLMM}}, \code{\link{MegaLMM_init}},
  \code{\link{print.MegaLMM_state}}, \code{\link{plot.MegaLMM_state}}
}
