% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_models.R
\name{voom_model}
\alias{voom_model}
\title{Sample Eta given a voom observation model}
\usage{
voom_model(observation_model_parameters, MegaLMM_state = list())
}
\arguments{
\item{observation_model_parameters}{List of parameters necessary for the data model.
Here, a Matrix of coordinates of NAs in Y}

\item{MegaLMM_state}{a MegaLMM_state object. Generally, only current_state and data_matrices is used. If
empty, will return a default set of parameters of the appropriate size for model initialization.}
}
\description{
\code{voom} or \code{voomWithQualityWeights} (limma) should be run on RNAseq data,
   resulting in logCPM (E) and inverseWeights (weights)
   for each observation. The observations (E) should be passed as Y, and the weights as
   observation_model_parameters$prec_Y.
}
\details{
When running \code{voom}, a fully-specified fixed effect model for the data should be specified.
}
\references{
Law, C. W., Chen, Y., Shi, W., & Smyth, G. K. (2014).
    voom: precision weights unlock linear model analysis tools for RNA-seq read counts.
    Genome Biology, 15(2), R29. http://doi.org/10.1186/gb-2004-5-10-r80
}
