% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yield_data}
\alias{yield_data}
\title{Yield BLUPs for 502 lines across 19 trials}
\format{
A data frame with 3318 rows and 4 columns
\describe{
  \item{Line}{Line identifier, corresponds to K}
  \item{Population}{Population identifier}
  \item{Env}{Trial identifier}
  \item{Yield}{Yield BLUP per Line:Env}
}
}
\usage{
yield_data
}
\description{
Yield BLUPs for 502 lines across 19 trials
}
\keyword{datasets}
