% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{regression_sampler_parallel}
\alias{regression_sampler_parallel}
\title{Draws samples from all ``fixed" coefficients (fixed and random) of a set of parallel linear regression models, conditional on the variance components.}
\usage{
regression_sampler_parallel(Y, W_base, W_list_, X, V_, h2s_index,
  chol_V_list_, Y_prec, Y_prec_a0, Y_prec_b0, prior_prec_alpha1,
  prior_prec_alpha2, prior_mean_beta, prior_prec_beta)
}
\arguments{
\item{Y}{n x p matrix of observations}

\item{W_base}{n x a1 matrix of W covariates common to all p. Can be NULL}

\item{W_list_}{p-list of n x a2 matrices of W covariates unique to each p. Can be NULL}

\item{X}{either X, a n x b matrix, or U, a n x m matrix. If U, then V must be non-NULL}

\item{V_}{m x b matrix if X is U, otherwise NULL}

\item{h2s_index}{p-vector of indices for to select appropriate V of each trait}

\item{chol_V_list_}{list of cholesky decompositions of V: RtR (each nxn). Can be either dense or sparse}

\item{Y_prec}{p-vector of Y current precisions}

\item{Y_prec_a0, Y_prec_b0}{scalars giving the shape and rate of the Gamma distribution for the prior on Y_prec}

\item{prior_prec_alpha1}{a1 x p matrix of prior precisions for alpha1}

\item{prior_prec_alpha2}{p-vector of precision of alpha2s for each trait}

\item{prior_mean_beta}{b x p matrix of prior means of beta}

\item{prior_prec_beta}{b x p matrix of prior precisions of beta}
}
\value{
List with elements: \itemize{
  \item alpha1 a1 x p matrix of alpha1
  \item alpha2 concatenated vector of alpha2 for all traits
  \item beta b x p matrix of beta
  \item Y_prec p x 1 vector of Y_prec
}
}
\description{
The model is either: \itemize{
\item y_i = W_base*alpha1 + W_list_[i]*alpha2 + X*beta + e, e ~ N(0,1/Y_prec[i]*V)
\item y_i = W_base*alpha1 + W_list_[i]*alpha2 + X*V_*beta + e, e ~ N(0,1/Y_prec[i]*V)
}
where \code{V = RtR}, priors on elements of alpha1, alpha2 and beta are independent.
Each column of Y is considered independent
}
